Refactored the project structure
