+++ Animations/status.py
@@ -6,7 +6,9 @@ def render(window, data):
     """
     window.erase()
     window.box()
-
+    
+    curses.curs_set(0)
+    curses.start_color()
     curses.init_pair(1, curses.COLOR_MAGENTA, curses.COLOR_BLACK)
     curses.init_pair(2, curses.COLOR_RED, curses.COLOR_BLACK)
     curses.init_pair(3, curses.COLOR_GREEN, curses.COLOR_BLACK)
